requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
celery[redis]
python-dotenv
httpx
google-api-python-client